// src/application/modules/wallet/dtos/get-coin-transactions.dto.ts
import type { CoinTransactionType } from "../../../../domain/entities/coin-transactions.entity";

export interface GetCoinTransactionsInput {
	userId: string;
	page: number;
	limit: number;
	sort?: "recent" | "old";
	type?: CoinTransactionType;
}

export interface CoinTransactionDto {
	id: string;
	amount: number;
	type: CoinTransactionType;
	transactionOwner: "platform" | "user";
	direction: "credit" | "debit";
	createdAt: Date;
}

export interface GetCoinTransactionsOutput {
	items: CoinTransactionDto[];
	total: number;
	page: number;
	limit: number;
	totalPages: number;
}